Extracted prompt utility method and SIMULATE flags
